requirements.txt
pandas
Pillow
keras
pickle-mixin
numpy
st_pages
seaborn
matplotlib
tensorflow
scikit-learn